jdk: Fix a whole bunch of warnings and a few build errors.
GCC is more strict than MSVC.

+++ openjdk/jdk/src/windows/native/sun/java2d/windows/GDIRenderer.cpp
@@ -320,7 +320,7 @@ Java_sun_java2d_windows_GDIRenderer_doDrawArc
         return;
     }
 
-    long sx, sy, ex, ey;
+    jint sx, sy, ex, ey;
     if (angleExtent >= 360 || angleExtent <= -360) {
         sx = ex = x + w;
         sy = ey = y + h/2;
@@ -593,8 +593,8 @@ Java_sun_java2d_windows_GDIRenderer_doFillArc
     if (wsdo == NULL) {
         return;
     }
-    long sx, sy, ex, ey;
-    int angleEnd;
+    jint sx, sy, ex, ey;
+    jint angleEnd;
     if (angleExtent < 0) {
         angleEnd = angleStart;
         angleStart += angleExtent;
